Use tokio::time::interval to improve sleeping
This changes the behaviour from sleeping a fixed amount of time _after_
the completion of the last run to sleeping the rest of the time
including the time it took to run.

In configurations where there is a _lot_ of messages per second the bot
might not be fast enough if it sleeps a minute after completion so this
should fix it.